modify gamesフォルダに移行
